1.5版本更新
1.宝宝入学 - 分享 增加复制链接
2.精品文章 - 分享 增加复制链接
3.我的收藏 - 列表 可以跳转到话题了
4.特长课程 - 分享 增加复制链接
5.优惠活动 - 分享 增加复制链接
6.话       题 - 分享 增加复制链接
7.芒果广告稳定性优化
8.新增webview保存图片、识别二维码功能
9.互       动 - 分享 增加分享链接
10.修改 主页 - getconfig 接口每天调用一次

+++ kindergartenApp/ShareActionSheet/HLActionSheet.h
@@ -7,8 +7,9 @@
 //
 
 #import <UIKit/UIKit.h>
+#import "UMSocialWechatHandler.h"
 
-typedef void(^ClickBlock)(int btnIndex);
+typedef void(^ClickBlock)(NSInteger btnIndex);
 typedef void(^CancelBlock)(void);
 
 @interface HLActionSheet : UIWindow
@@ -16,5 +17,7 @@ typedef void(^CancelBlock)(void);
 - (instancetype)initWithTitles:(NSArray *)titles iconNames:(NSArray *)iconNames;
 
 - (void)showActionSheetWithClickBlock:(ClickBlock)clickBlock cancelBlock:(CancelBlock)cancelBlock;
+
+
 
 @end
